+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 = 14.14 - 2025-xx-xx
 - **New:** Added an option to to disable word count analytics in reports.
+- **New**: Add display dynamic promo banner.
 
 = 14.13.5 - 2025-05-x =
 - **Fix:** Fixed issue where page views were zero when using a taxonomy filter other than 'category'.